#30 Unit testing for Event and Calendar Item

+++ web-calendar-api/WebCalendar.Data/Repositories/EventRepository.cs
@@ -19,6 +19,10 @@ namespace WebCalendar.Data.Repositories
     public Event GetEvent(int id)
     {
       var @event =_context.Events.Find(id);
+      if (@event == null)
+      {
+        return null;
+      }
       _context.Entry(@event).State = EntityState.Detached;
       return @event;
     }
@@ -64,13 +68,13 @@ namespace WebCalendar.Data.Repositories
        })
        .FirstOrDefault();
 
-    public void AddSeriesOfCalendarEvents(IEnumerable<Event> calendarEvents, int? seriesId)
+    public void AddSeriesOfCalendarEvents(IEnumerable<Event> calendarEvents)
     {
       _context.Events.AddRange(calendarEvents);
       _context.SaveChanges();
     }
 
-    public Event AddCalendarEvents(Event calendarEvent)
+    public Event AddCalendarEvent(Event calendarEvent)
     {
       _context.Events.Add(calendarEvent);
       _context.SaveChanges();
@@ -78,15 +82,13 @@ namespace WebCalendar.Data.Repositories
       return calendarEvent;
     }
 
-    public void UpdateEvent(Event calendarEvent)
-    {
-      _context.Events.Update(calendarEvent);
-      _context.SaveChanges();
-    }
-
     public Event DeleteCalendarEvent(int calendarEventId)
     {
       var currentEvent = _context.Events.Find(calendarEventId);
+      if (currentEvent == null)
+      {
+        return null;
+      }
       _context.Events.Remove(currentEvent);
       _context.SaveChanges();
       return currentEvent;
